:sparkles: 결제 주문 연결

+++ src/main/java/com/t2m/g2nee/front/order/dto/request/OrderForm.java
@@ -63,7 +63,7 @@ public class OrderForm {
 
     private Long couponId;//null가능
 
-    private Long point;//null가능
+    private Integer point;//null가능
 }
